+++ unity/LoboBranco/Assets/_Project/Code/Player/PlayerInputReader.cs
@@ -38,6 +38,8 @@ namespace LoboBranco.Player
         InputAction _interact;
         InputAction _witcherSenses;
         InputAction _cycleStance;
+        InputAction _switchSteel;
+        InputAction _switchSilver;
 
         // ---------------------------------------------------------------- estado
 
@@ -57,6 +59,12 @@ namespace LoboBranco.Player
 
         /// <summary>Roda de postura. Positivo avanca, negativo volta (docs/02 secao 4).</summary>
         public event Action<int> StanceCycled;
+
+        /// <summary>Pedido de espada de aco. Quem decide se da e a maquina de estados.</summary>
+        public event Action SwitchSteelPressed;
+
+        /// <summary>Pedido de espada de prata.</summary>
+        public event Action SwitchSilverPressed;
 
         // ---------------------------------------------------------------- ciclo
 
@@ -82,6 +90,8 @@ namespace LoboBranco.Player
             _interact = _map.FindAction("Interact", true);
             _witcherSenses = _map.FindAction("WitcherSenses", true);
             _cycleStance = _map.FindAction("CycleStance", true);
+            _switchSteel = _map.FindAction("SwitchSteel", true);
+            _switchSilver = _map.FindAction("SwitchSilver", true);
         }
 
         void OnEnable()
@@ -95,6 +105,8 @@ namespace LoboBranco.Player
             _interact.performed += OnInteract;
             _witcherSenses.performed += OnWitcherSenses;
             _cycleStance.performed += OnCycleStance;
+            _switchSteel.performed += OnSwitchSteel;
+            _switchSilver.performed += OnSwitchSilver;
 
             _map.Enable();
         }
@@ -110,6 +122,8 @@ namespace LoboBranco.Player
             _interact.performed -= OnInteract;
             _witcherSenses.performed -= OnWitcherSenses;
             _cycleStance.performed -= OnCycleStance;
+            _switchSteel.performed -= OnSwitchSteel;
+            _switchSilver.performed -= OnSwitchSilver;
 
             _map.Disable();
 
@@ -135,6 +149,8 @@ namespace LoboBranco.Player
         void OnCastSign(InputAction.CallbackContext _) => CastSignPressed?.Invoke();
         void OnInteract(InputAction.CallbackContext _) => InteractPressed?.Invoke();
         void OnWitcherSenses(InputAction.CallbackContext _) => WitcherSensesPressed?.Invoke();
+        void OnSwitchSteel(InputAction.CallbackContext _) => SwitchSteelPressed?.Invoke();
+        void OnSwitchSilver(InputAction.CallbackContext _) => SwitchSilverPressed?.Invoke();
 
         /// <summary>
         /// A acao e um eixo, e nao um botao, porque roda de mouse e direcional sao eixos.